+++ App.tsx
@@ -3,14 +3,14 @@ import { createNativeStackNavigator } from "@react-navigation/native-stack";
 import { QueryClient, QueryClientProvider } from "@tanstack/react-query";
 import { atom } from "jotai";
 import { SafeAreaView } from "react-native";
-import Notifications from "./containers/NotificationsContainer/Notifications";
+import Notifications from "./components/notifications/Notifications";
 import BottomNavigation from "./navigation/BottomNavigation";
-import ChangePasswordScreen from "./screens/auth/ChangePasswordScreen";
-import LoginScreen from "./screens/auth/LoginScreen";
-import RegisterScreen from "./screens/auth/RegisterScreen";
-import ForgotPasswordCodeScreen from "./screens/auth/forgot-password/ForgotPasswordCodeScreen";
-import ForgotPasswordNewPasswordScreen from "./screens/auth/forgot-password/ForgotPasswordNewPasswordScreen";
-import ForgotPasswordScreen from "./screens/auth/forgot-password/ForgotPasswordScreen";
+import ChangePasswordScreen from "./screens/common/auth/ChangePasswordScreen";
+import LoginScreen from "./screens/common/auth/LoginScreen";
+import RegisterScreen from "./screens/common/auth/RegisterScreen";
+import ForgotPasswordCodeScreen from "./screens/common/auth/forgot-password/ForgotPasswordCodeScreen";
+import ForgotPasswordNewPasswordScreen from "./screens/common/auth/forgot-password/ForgotPasswordNewPasswordScreen";
+import ForgotPasswordScreen from "./screens/common/auth/forgot-password/ForgotPasswordScreen";
 import LandingScreen from "./screens/common/LandingScreen";
 import { CustomErrorBoundary } from "./screens/errors/ErrorBoundary";
