+++ components/chat/WritingAnimation.tsx
