+++ components/common/PrimaryButton.tsx
@@ -1,19 +1,28 @@
-import { Platform, Pressable, StyleSheet, Text, View } from "react-native";
-import Colors from "../theme/colors";
+import {
+  ActivityIndicator,
+  Platform,
+  Pressable,
+  StyleSheet,
+  Text,
+  View,
+} from "react-native";
+import Colors from "../../theme/colors";
 
-interface SecondaryButtonProps {
+interface PrimaryButtonProps {
   children: React.ReactNode;
+  loading?: boolean;
   onPress?: () => void;
 }
 
-const SecondaryButton = (props: SecondaryButtonProps) => {
-  const { children, onPress } = props;
+const PrimaryButton = (props: PrimaryButtonProps) => {
+  const { children, onPress, loading } = props;
 
   return (
     <View style={styles.outerContainer}>
       <Pressable
         onPress={onPress}
-        android_ripple={{ color: Colors.gray[100] }}
+        disabled={loading}
+        android_ripple={{ color: Colors.primary[600] }}
         style={({ pressed }) => {
           if (Platform.OS === "ios") {
             return [styles.innerContainer, pressed && styles.pressed];
@@ -21,7 +30,11 @@ const SecondaryButton = (props: SecondaryButtonProps) => {
           return styles.innerContainer;
         }}
       >
-        <Text style={styles.buttonText}>{children}</Text>
+        {loading ? (
+          <ActivityIndicator color="white" />
+        ) : (
+          <Text style={styles.buttonText}>{children}</Text>
+        )}
       </Pressable>
     </View>
   );
@@ -30,10 +43,8 @@ const SecondaryButton = (props: SecondaryButtonProps) => {
 const styles = StyleSheet.create({
   outerContainer: {
     borderRadius: 8,
-    borderWidth: 1.5,
     marginVertical: 8,
     overflow: "hidden",
-    borderColor: Colors.primary[500],
     shadowColor: Colors.gray[900],
     shadowOpacity: 0.5,
     shadowRadius: 8,
@@ -44,19 +55,27 @@ const styles = StyleSheet.create({
     elevation: 12,
   },
   innerContainer: {
-    backgroundColor: "white",
+    backgroundColor: Colors.primary[500],
     paddingVertical: 16,
     paddingHorizontal: 32,
+    elevation: 4,
   },
   buttonText: {
-    color: Colors.primary[500],
+    color: "white",
     fontSize: 18,
     fontWeight: "bold",
     textAlign: "center",
+    shadowColor: "black",
+    shadowOpacity: 0.5,
+    shadowRadius: 8,
+    shadowOffset: {
+      width: 0,
+      height: 4,
+    },
   },
   pressed: {
     opacity: 0.75,
   },
 });
 
-export default SecondaryButton;
+export default PrimaryButton;
