+++ components/common/password/StrengthMeter.tsx
