+++ components/common/password/index.tsx
