+++ screens/WordListsScreen.tsx
@@ -1,0 +1,5 @@
+const WordListsScreen = () => {
+  return <div></div>;
+};
+
+export default WordListsScreen;
