+++ screens/common/LandingScreen.tsx
@@ -1,9 +1,9 @@
 import { useMutation } from "@tanstack/react-query";
 import { useEffect } from "react";
 import { StyleSheet, Text, View } from "react-native";
-import PrimaryButton from "../../components/PrimaryButton";
-import SecondaryButton from "../../components/SecondaryButton";
-import Title from "../../components/Title";
+import PrimaryButton from "../../components/common/PrimaryButton";
+import SecondaryButton from "../../components/common/SecondaryButton";
+import Title from "../../components/common/Title";
 import useNotifications from "../../hooks/useNotifications";
 import { checkAuth } from "../../services/auth";
 import Colors from "../../theme/colors";
