+++ screens/common/auth/ChangePasswordScreen.tsx
@@ -6,15 +6,15 @@ import {
   useForm,
 } from "react-hook-form";
 import { ScrollView, StyleSheet, Text, View } from "react-native";
-import PrimaryButton from "../../components/PrimaryButton";
-import EmailTextInput from "../../components/input/EmailTextInput";
-import PasswordTextInput from "../../components/input/PasswordTextInput";
-import PasswordInputWithRequirements from "../../containers/PasswordInputWithRequirements/PasswordInputWithRequirements";
-import { Requirement } from "../../containers/PasswordInputWithRequirements/Requirement";
-import useNotifications from "../../hooks/useNotifications";
-import { changePassword, register } from "../../services/auth/Auth.service";
-import Colors from "../../theme/colors";
-import { generateErrorResponseMessage } from "../../utils/httpUtils";
+import PrimaryButton from "../../../components/common/PrimaryButton";
+import EmailTextInput from "../../../components/common/input/EmailTextInput";
+import PasswordTextInput from "../../../components/common/input/PasswordTextInput";
+import PasswordInputWithRequirements from "../../../components/common/password/PasswordInputWithRequirements";
+import { Requirement } from "../../../components/common/password/Requirement";
+import useNotifications from "../../../hooks/useNotifications";
+import { changePassword, register } from "../../../services/auth/Auth.service";
+import Colors from "../../../theme/colors";
+import { generateErrorResponseMessage } from "../../../utils/httpUtils";
 
 type ChangePasswordFormValues = {
   oldPassword: string;
