+++ screens/common/auth/LoginScreen.tsx
@@ -6,16 +6,16 @@ import {
   useForm,
 } from "react-hook-form";
 import { ScrollView, StyleSheet, Text, View } from "react-native";
-import PrimaryButton from "../../components/PrimaryButton";
-import EmailTextInput from "../../components/input/EmailTextInput";
-import PasswordTextInput from "../../components/input/PasswordTextInput";
-import PrimaryTextInput from "../../components/input/PrimaryTextInput";
-import useUser from "../../hooks/auth/useUser";
-import useNotifications from "../../hooks/useNotifications";
-import { login } from "../../services/auth";
-import Colors from "../../theme/colors";
-import { APIResponse } from "../../types/common";
-import { generateErrorResponseMessage } from "../../utils/httpUtils";
+import PrimaryButton from "../../../components/common/PrimaryButton";
+import EmailTextInput from "../../../components/common/input/EmailTextInput";
+import PasswordTextInput from "../../../components/common/input/PasswordTextInput";
+import PrimaryTextInput from "../../../components/common/input/PrimaryTextInput";
+import useUser from "../../../hooks/auth/useUser";
+import useNotifications from "../../../hooks/useNotifications";
+import { login } from "../../../services/auth";
+import Colors from "../../../theme/colors";
+import { APIResponse } from "../../../types/common";
+import { generateErrorResponseMessage } from "../../../utils/httpUtils";
 
 type FormValues = {
   email: string;
