+++ screens/common/auth/RegisterScreen.tsx
@@ -1,15 +1,15 @@
 import { useMutation } from "@tanstack/react-query";
 import { FormProvider, useForm } from "react-hook-form";
 import { ScrollView, StyleSheet, View } from "react-native";
-import PrimaryButton from "../../components/PrimaryButton";
-import EmailTextInput from "../../components/input/EmailTextInput";
-import PasswordTextInput from "../../components/input/PasswordTextInput";
-import PrimaryTextInput from "../../components/input/PrimaryTextInput";
-import PasswordInputWithRequirements from "../../containers/PasswordInputWithRequirements/PasswordInputWithRequirements";
-import { Requirement } from "../../containers/PasswordInputWithRequirements/Requirement";
-import useNotifications from "../../hooks/useNotifications";
-import { login, register } from "../../services/auth";
-import { generateErrorResponseMessage } from "../../utils/httpUtils";
+import PrimaryButton from "../../../components/common/PrimaryButton";
+import EmailTextInput from "../../../components/common/input/EmailTextInput";
+import PasswordTextInput from "../../../components/common/input/PasswordTextInput";
+import PrimaryTextInput from "../../../components/common/input/PrimaryTextInput";
+import PasswordInputWithRequirements from "../../../components/common/password/PasswordInputWithRequirements";
+import { Requirement } from "../../../components/common/password/Requirement";
+import useNotifications from "../../../hooks/useNotifications";
+import { login, register } from "../../../services/auth";
+import { generateErrorResponseMessage } from "../../../utils/httpUtils";
 
 type RegisterFormValues = {
   userName: string;
