+++ screens/common/auth/forgot-password/ForgotPasswordNewPasswordScreen.tsx
@@ -6,17 +6,17 @@ import {
   useForm,
 } from "react-hook-form";
 import { ScrollView, StyleSheet, Text, View } from "react-native";
-import PrimaryButton from "../../../components/PrimaryButton";
-import PasswordTextInput from "../../../components/input/PasswordTextInput";
-import PasswordInputWithRequirements from "../../../containers/PasswordInputWithRequirements/PasswordInputWithRequirements";
-import { Requirement } from "../../../containers/PasswordInputWithRequirements/Requirement";
-import useNotifications from "../../../hooks/useNotifications";
+import PrimaryButton from "../../../../components/common/PrimaryButton";
+import PasswordTextInput from "../../../../components/common/input/PasswordTextInput";
+import PasswordInputWithRequirements from "../../../../components/common/password/PasswordInputWithRequirements";
+import { Requirement } from "../../../../components/common/password/Requirement";
+import useNotifications from "../../../../hooks/useNotifications";
 import {
   changePassword,
   register,
   saveResetPassword,
-} from "../../../services/auth/Auth.service";
-import { generateErrorResponseMessage } from "../../../utils/httpUtils";
+} from "../../../../services/auth/Auth.service";
+import { generateErrorResponseMessage } from "../../../../utils/httpUtils";
 
 type ForgotPasswordNewPasswordFormValues = {
   password: string;
