+++ screens/user/ProfileScreen.tsx
@@ -8,7 +8,7 @@ import {
   TouchableWithoutFeedback,
   View,
 } from "react-native";
-import SecondaryButton from "../../components/SecondaryButton";
+import SecondaryButton from "../../components/common/SecondaryButton";
 import Colors from "../../theme/colors";
 
 const avatarPlaceholderImg = require("../../assets/profile-default.jpg");
